+++ src/credential.rs
@@ -1,6 +1,6 @@
 /*!
 
-# Platorm-independent secure storage model
+# Platform-independent secure storage model
 
 This module defines a plug and play model for platform-specific credential stores.
 The model comprises two traits: [CredentialBuilderApi] for the underlying store
@@ -28,6 +28,7 @@ if  matches!(persistence, credential::CredentialPersistence::UntilDelete) {
 ```
  */
 use std::any::Any;
+use std::collections::HashMap;
 
 use super::Result;
 
@@ -43,15 +44,52 @@ pub trait CredentialApi {
     /// This will persist the secret in the underlying store.
     fn set_secret(&self, password: &[u8]) -> Result<()>;
 
-    /// Retrieve a password (a string) from the credential, if one has been set.
+    /// Retrieve the password (a string) from the underlying credential.
     ///
-    /// This has no effect on the underlying store.
+    /// This has no effect on the underlying store. If there is no credential
+    /// for this entry, a [NoEntry](crate::Error::NoEntry) error is returned.
     fn get_password(&self) -> Result<String>;
 
-    /// Retrieve a secret (a byte array) from the credential, if one has been set.
+    /// Retrieve a secret (a byte array) from the credential.
     ///
-    /// This has no effect on the underlying store.
+    /// This has no effect on the underlying store. If there is no credential
+    /// for this entry, a [NoEntry](crate::Error::NoEntry) error is returned.
     fn get_secret(&self) -> Result<Vec<u8>>;
+
+    /// Get the secure store attributes on this entry's credential.
+    ///
+    /// Each credential store may support reading and updating different
+    /// named attributes; see the documentation on each of the stores
+    /// for details. Note that the keyring itself uses some of these
+    /// attributes to map entries to their underlying credential; these
+    /// _controlled_ attributes are not available for reading or updating.
+    ///
+    /// We provide a default (no-op) implementation of this method
+    /// for backward compatibility with stores that don't implement it.
+    fn get_attributes(&self) -> Result<HashMap<String, String>> {
+        // this should err in the same cases as get_secret, so first call that for effect
+        self.get_secret()?;
+        // if we got this far, return success with no attributes
+        Ok(HashMap::new())
+    }
+
+    /// Update the secure store attributes on this entry's credential.
+    ///
+    /// Each credential store may support reading and updating different
+    /// named attributes; see the documentation on each of the stores
+    /// for details. The implementation will ignore any attribute names
+    /// that you supply that are not available for update. Because the
+    /// names used by the different stores tend to be distinct, you can
+    /// write cross-platform code that will work correctly on each platform.
+    ///
+    /// We provide a default no-op implementation of this method
+    /// for backward compatibility with stores that don't implement it.
+    fn update_attributes(&self, _: &HashMap<&str, &str>) -> Result<()> {
+        // this should err in the same cases as get_secret, so first call that for effect
+        self.get_secret()?;
+        // if we got this far, return success after setting no attributes
+        Ok(())
+    }
 
     /// Delete the underlying credential, if there is one.
     ///
@@ -112,6 +150,7 @@ pub trait CredentialBuilderApi {
     /// This typically has no effect on the content of the underlying store.
     /// A credential need not be persisted until its password is set.
     fn build(&self, target: Option<&str>, service: &str, user: &str) -> Result<Box<Credential>>;
+
     /// Return the underlying concrete object cast to [Any].
     ///
     /// Because credential builders need not have any internal structure,